Rename FailureInjector to TestingFailureInjector
